fix: carga dinámica en lucide icons eliminada
