--- src/components/LucideIcon.tsx
@@ -1,15 +1,0 @@
-import dynamic from 'next/dynamic'
-import { LucideProps } from 'lucide-react'
-import dynamicIconImports from 'lucide-react/dynamicIconImports'
-
-interface IconProps extends LucideProps {
-  name: keyof typeof dynamicIconImports
-}
-
-const Icon = ({ name, ...props }: IconProps) => {
-  const LucideIcon = dynamic(dynamicIconImports[name])
-
-  return <LucideIcon {...props} />
-}
-
-export default Icon
