+++ src/types/menuTypes.ts
@@ -1,10 +1,10 @@
-import dynamicIconImports from 'lucide-react/dynamicIconImports'
+import { ReactNode } from 'react'
 
 export interface MenuItems {
   id: string
   title: string
   type: 'group' | 'item' | 'collapse'
-  icon?: keyof typeof dynamicIconImports
+  icon?: ReactNode
   url?: string
   target?: boolean
   children?: MenuItems[]
